# openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Mydata API",
		"version": "1.0",
		"description": "API for handling electronic signatures, certification operations, and account information"
	},
	"servers": [
		{
			"url": "http://localhost:5000",
			"description": "Local development server"
		},
		{
			"url": "https://localhost:4200",
			"description": "Other bank API server"
		}
	],
...
	"paths": {
...
		"/api/ca/sign_request": {
			"post": {
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Handles electronic signature requests.",
				"description": "Validates the request, processes the electronic signature, and returns a certification transaction ID.",
				"tags": ["Certification Authority"],
				"operationId": "getIA102",
				"parameters": [
					{
						"in": "header",
						"name": "x-api-tran-id",
						"schema": {
							"type": "string",
							"maxLength": 25,
							"pattern": "^[a-zA-Z0-9]{1,25}$",
							"example": "anya123456SDFRES202503051"
						},
						"required": true,
						"description": "API transaction ID for request tracking"
					},
					{
						"in": "header",
						"name": "Authorization",
						"schema": {
							"type": "string",
							"example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
						},
						"required": true,
						"description": "Bearer token for authorization"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Electronic signature request successful.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SignResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request parameters.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized request.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Invalid token.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "JWT Authorization header using the Bearer scheme"
			}
		},
		"schemas": {
			"SignRequest": {
				"type": "object",
				"required": ["sign_tx_id", "user_ci", "real_name", "phone_num", "request_title", "consent_list"],
				"properties": {
					"sign_tx_id": {
						"type": "string",
						"description": "Transaction ID for signing",
						"example": "anya123456_certauth00_20250305120000_anyaserial00"
					},
					"user_ci": {
						"type": "string",
						"description": "Base64 encoded user identification",
						"example": "RzFHZmwzUnQyWHM="
					},
					"real_name": {
						"type": "string",
						"description": "User's full name",
						"example": "John Doe"
					},
					"phone_num": {
						"type": "string",
						"description": "User's phone number",
						"example": "+821012345678"
					},
					"request_title": {
						"type": "string",
						"description": "Title of the request",
						"example": "Request for Consent to Use Personal Information"
					},
					"device_code": {
						"type": "string",
						"description": "Device code (PC, MO, TB)",
						"example": "PC"
					},
					"device_browser": {
						"type": "string",
						"description": "Browser code",
						"example": "WB"
					},
					"return_app_scheme_url": {
						"type": "string",
						"description": "URL to return to after signing",
						"example": "https://anya-bank.com/return"
					},
					"consent_type": {
						"type": "string",
						"description": "Type of consent",
						"example": "1"
					},
					"consent_cnt": {
						"type": "integer",
						"description": "Number of consents",
						"example": 3
					},
					"consent_list": {
						"type": "array",
						"description": "List of consents",
						"items": {
							"$ref": "#/components/schemas/Consent"
						}
					}
				}
			},
			"Consent": {
				"type": "object",
				"required": ["tx_id", "consent_title", "consent", "consent_len"],
				"properties": {
					"tx_id": {
						"type": "string",
						"description": "Transaction ID for the consent",
						"example": "MD_anya123456_bond123456_RA20215001_certauth00_20250305120000_XXAB00490001"
					},
					"consent_title": {
						"type": "string",
						"description": "Title of the consent",
						"example": "Consent Request for Transmission"
					},
					"consent": {
						"type": "string",
						"description": "Base64 encoded consent content",
						"example": "Y29uc2VudC0wMDE="
					},
					"consent_len": {
						"type": "integer",
						"description": "Length of the consent content",
						"example": 12
					}
				}
			},
			"SignResponse": {
				"type": "object",
				"required": ["rsp_code", "rsp_msg", "cert_tx_id"],
				"properties": {
					"rsp_code": {
						"type": "string",
						"description": "Response code",
						"example": "0000"
					},
					"rsp_msg": {
						"type": "string",
						"description": "Response message",
						"example": "Electronic signature request successful, cert_tx_id has been provided."
					},
					"sign_ios_app_scheme_url": {
						"type": "string",
						"description": "iOS app URL scheme for signing",
						"example": "mydataauth://auth?tx_id=ORG2025001"
					},
					"sign_aos_app_scheme_url": {
						"type": "string",
						"description": "Android app URL scheme for signing",
						"example": "mydataauth://auth?tx_id=ORG2025001"
					},
					"sign_web_url": {
						"type": "string",
						"description": "Web URL for signing",
						"example": "https://mydataauth.com/auth?tx_id=ORG2025001"
					},
					"cert_tx_id": {
						"type": "string",
						"description": "Certificate transaction ID",
						"example": "CERT2025001"
					}
				}
			},
...
			"ErrorResponse": {
				"type": "object",
				"required": ["code", "message"],
				"properties": {
					"code": {
						"type": "string",
						"description": "Error code",
						"example": "ERR_400"
					},
					"message": {
						"type": "string",
						"description": "Error message",
						"example": "Invalid request parameters"
					}
				}
			}
		}
	}
}